# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Blockchain-based certificate issuance and verification system",
        "title": "Certificate System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/university/register": {
            "post": {
                "description": "Register a university with domain and generate Cardano wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "university"
                ],
                "summary": "Register a new university",
                "parameters": [
                    {
                        "description": "University registration details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/university/{id}/domain-proof": {
            "get": {
                "description": "Get the verification file content for domain proof",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "university"
                ],
                "summary": "Get domain verification proof",
                "parameters": [
                    {
                        "type": "string",
                        "description": "University ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.DomainVerificationResponse"
                        }
                    }
                }
            }
        },
        "/university/{id}/verify-domain": {
            "post": {
                "description": "Verify that the university owns the domain by checking the verification file",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "university"
                ],
                "summary": "Verify domain ownership",
                "parameters": [
                    {
                        "type": "string",
                        "description": "University ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/verify/blockchain/{certID}": {
            "get": {
                "description": "Get blockchain explorer links for a certificate",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "verification"
                ],
                "summary": "Get blockchain information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Certificate ID",
                        "name": "certID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.DomainVerificationResponse": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string"
                },
                "file_content": {
                    "type": "object",
                    "additionalProperties": true
                },
                "file_path": {
                    "type": "string"
                },
                "instructions": {
                    "type": "string"
                },
                "university_id": {
                    "type": "string"
                }
            }
        },
...
        "handlers.RegisterRequest": {
            "type": "object",
            "required": [
                "domain",
                "email",
                "name",
                "password"
            ],
            "properties": {
                "domain": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 12
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}